+++ aclk/https_client.c
@@ -4,19 +4,11 @@
 
 #include "https_client.h"
 
-#include "mqtt_websockets/c-rbuf/include/ringbuffer.h"
-
 #include "aclk_util.h"
 
 #include "daemon/global_statistics.h"
 
 #define DEFAULT_CHUNKED_RESPONSE_BUFFER_SIZE (4096)
-
-enum http_parse_state {
-    HTTP_PARSE_INITIAL = 0,
-    HTTP_PARSE_HEADERS,
-    HTTP_PARSE_CONTENT
-};
 
 static const char *http_req_type_to_str(http_req_type_t req) {
     switch (req) {
@@ -33,27 +25,6 @@ static const char *http_req_type_to_str(http_req_type_t req) {
 
 #define TRANSFER_ENCODING_CHUNKED (-2)
 
-typedef struct {
-    enum http_parse_state state;
-    int content_length;
-    int http_code;
-
-    // for chunked data only
-    char *chunked_response;
-    size_t chunked_response_size;
-    size_t chunked_response_written;
-
-    enum chunked_content_state {
-        CHUNKED_CONTENT_CHUNK_SIZE = 0,
-        CHUNKED_CONTENT_CHUNK_DATA,
-        CHUNKED_CONTENT_CHUNK_END_CRLF,
-        CHUNKED_CONTENT_FINAL_CRLF
-    } chunked_content_state;
-
-    size_t chunk_size;
-    size_t chunk_got;
-} http_parse_ctx;
-
 #define HTTP_PARSE_CTX_INITIALIZER { .state = HTTP_PARSE_INITIAL, .content_length = -1, .http_code = 0 }
 static inline void http_parse_ctx_clear(http_parse_ctx *ctx) {
     ctx->state = HTTP_PARSE_INITIAL;
@@ -63,9 +34,6 @@ static inline void http_parse_ctx_clear(http_parse_ctx *ctx) {
 
 #define POLL_TO_MS 100
 
-#define NEED_MORE_DATA  0
-#define PARSE_SUCCESS   1
-#define PARSE_ERROR    -1
 #define HTTP_LINE_TERM "\x0D\x0A"
 #define RESP_PROTO "HTTP/1.1 "
 #define HTTP_KEYVAL_SEPARATOR ": "
@@ -169,8 +137,8 @@ static int process_chunked_content(rbuf_t buf, http_parse_ctx *parse_ctx)
             case CHUNKED_CONTENT_CHUNK_SIZE:
                 if (!rbuf_find_bytes(buf, HTTP_LINE_TERM, strlen(HTTP_LINE_TERM), &idx)) {
                     if (rbuf_bytes_available(buf) >= rbuf_get_capacity(buf))
-                        return PARSE_ERROR;
-                    return NEED_MORE_DATA;
+                        return HTTP_PARSE_ERROR;
+                    return HTTP_PARSE_NEED_MORE_DATA;
                 }
                 if (idx == 0) {
                     parse_ctx->chunked_content_state = CHUNKED_CONTENT_FINAL_CRLF;
@@ -178,7 +146,7 @@ static int process_chunked_content(rbuf_t buf, http_parse_ctx *parse_ctx)
                 }
                 if (idx >= HTTP_HDR_BUFFER_SIZE) {
                     netdata_log_error("Chunk size is too long");
-                    return PARSE_ERROR;
+                    return HTTP_PARSE_ERROR;
                 }
                 char buf_size[HTTP_HDR_BUFFER_SIZE];
                 rbuf_pop(buf, buf_size, idx);
@@ -186,13 +154,13 @@ static int process_chunked_content(rbuf_t buf, http_parse_ctx *parse_ctx)
                 long chunk_size = strtol(buf_size, NULL, 16);
                 if (chunk_size < 0 || chunk_size == LONG_MAX) {
                     netdata_log_error("Chunk size out of range");
-                    return PARSE_ERROR;
+                    return HTTP_PARSE_ERROR;
                 }
                 parse_ctx->chunk_size = chunk_size;
                 if (parse_ctx->chunk_size == 0) {
                     if (errno == EINVAL) {
                         netdata_log_error("Invalid chunk size");
-                        return PARSE_ERROR;
+                        return HTTP_PARSE_ERROR;
                     }
                     parse_ctx->chunked_content_state = CHUNKED_CONTENT_CHUNK_END_CRLF;
                     continue;
@@ -204,7 +172,7 @@ static int process_chunked_content(rbuf_t buf, http_parse_ctx *parse_ctx)
                 // fallthrough
             case CHUNKED_CONTENT_CHUNK_DATA:
                 if (!(bytes_to_copy = rbuf_bytes_available(buf)))
-                    return NEED_MORE_DATA;
+                    return HTTP_PARSE_NEED_MORE_DATA;
                 if (bytes_to_copy > parse_ctx->chunk_size - parse_ctx->chunk_got)
                     bytes_to_copy = parse_ctx->chunk_size - parse_ctx->chunk_got;
                 rbuf_pop(buf, parse_ctx->chunked_response + parse_ctx->chunked_response_written, bytes_to_copy);
@@ -217,19 +185,19 @@ static int process_chunked_content(rbuf_t buf, http_parse_ctx *parse_ctx)
             case CHUNKED_CONTENT_FINAL_CRLF:
             case CHUNKED_CONTENT_CHUNK_END_CRLF:
                 if (rbuf_bytes_available(buf) < strlen(HTTP_LINE_TERM))
-                    return NEED_MORE_DATA;
+                    return HTTP_PARSE_NEED_MORE_DATA;
                 char buf_crlf[strlen(HTTP_LINE_TERM)];
                 rbuf_pop(buf, buf_crlf, strlen(HTTP_LINE_TERM));
                 if (memcmp(buf_crlf, HTTP_LINE_TERM, strlen(HTTP_LINE_TERM))) {
                     netdata_log_error("CRLF expected");
-                    return PARSE_ERROR;
+                    return HTTP_PARSE_ERROR;
                 }
                 if (parse_ctx->chunked_content_state == CHUNKED_CONTENT_FINAL_CRLF) {
                     if (parse_ctx->chunked_response_size != parse_ctx->chunked_response_written)
                         netdata_log_error("Chunked response size mismatch");
                     chunked_response_buffer_grow_by(parse_ctx, 1);
                     parse_ctx->chunked_response[parse_ctx->chunked_response_written] = 0;
-                    return PARSE_SUCCESS;
+                    return HTTP_PARSE_SUCCESS;
                 }
                 if (parse_ctx->chunk_size == 0) {
                     parse_ctx->chunked_content_state = CHUNKED_CONTENT_FINAL_CRLF;
@@ -241,34 +209,34 @@ static int process_chunked_content(rbuf_t buf, http_parse_ctx *parse_ctx)
     } while(1);
 }
 
-static int parse_http_response(rbuf_t buf, http_parse_ctx *parse_ctx)
+int parse_http_response(rbuf_t buf, http_parse_ctx *parse_ctx)
 {
     int idx;
     char rc[4];
 
     do {
         if (parse_ctx->state != HTTP_PARSE_CONTENT && !rbuf_find_bytes(buf, HTTP_LINE_TERM, strlen(HTTP_LINE_TERM), &idx))
-            return NEED_MORE_DATA;
+            return HTTP_PARSE_NEED_MORE_DATA;
         switch (parse_ctx->state) {
             case HTTP_PARSE_INITIAL:
                 if (rbuf_memcmp_n(buf, RESP_PROTO, strlen(RESP_PROTO))) {
                     netdata_log_error("Expected response to start with \"%s\"", RESP_PROTO);
-                    return PARSE_ERROR;
+                    return HTTP_PARSE_ERROR;
                 }
                 rbuf_bump_tail(buf, strlen(RESP_PROTO));
                 if (rbuf_pop(buf, rc, 4) != 4) {
                     netdata_log_error("Expected HTTP status code");
-                    return PARSE_ERROR;
+                    return HTTP_PARSE_ERROR;
                 }
                 if (rc[3] != ' ') {
                     netdata_log_error("Expected space after HTTP return code");
-                    return PARSE_ERROR;
+                    return HTTP_PARSE_ERROR;
                 }
                 rc[3] = 0;
                 parse_ctx->http_code = atoi(rc);
                 if (parse_ctx->http_code < 100 || parse_ctx->http_code >= 600) {
                     netdata_log_error("HTTP code not in range 100 to 599");
-                    return PARSE_ERROR;
+                    return HTTP_PARSE_ERROR;
                 }
 
                 rbuf_find_bytes(buf, HTTP_LINE_TERM, strlen(HTTP_LINE_TERM), &idx);
@@ -284,7 +252,7 @@ static int parse_http_response(rbuf_t buf, http_parse_ctx *parse_ctx)
                     break;
                 }
                 if (parse_http_hdr(buf, parse_ctx))
-                    return PARSE_ERROR;
+                    return HTTP_PARSE_ERROR;
                 rbuf_find_bytes(buf, HTTP_LINE_TERM, strlen(HTTP_LINE_TERM), &idx);
                 rbuf_bump_tail(buf, idx + strlen(HTTP_LINE_TERM));
                 break;
@@ -294,11 +262,11 @@ static int parse_http_response(rbuf_t buf, http_parse_ctx *parse_ctx)
                     return process_chunked_content(buf, parse_ctx);
 
                 if (parse_ctx->content_length < 0)
-                    return PARSE_SUCCESS;
+                    return HTTP_PARSE_SUCCESS;
 
                 if (rbuf_bytes_available(buf) >= (size_t)parse_ctx->content_length)
-                    return PARSE_SUCCESS;
-                return NEED_MORE_DATA;
+                    return HTTP_PARSE_SUCCESS;
+                return HTTP_PARSE_NEED_MORE_DATA;
         }
     } while(1);
 }
@@ -486,7 +454,7 @@ static int read_parse_response(https_req_ctx_t *ctx) {
         } while (ctx->poll_fd.events == 0 && rbuf_bytes_free(ctx->buf_rx) > 0);
     } while (!(ret = parse_http_response(ctx->buf_rx, &ctx->parse_ctx)));
 
-    if (ret != PARSE_SUCCESS) {
+    if (ret != HTTP_PARSE_SUCCESS) {
         netdata_log_error("Error parsing HTTP response");
         return 1;
     }
